+++ rllm/engine/rollout/__init__.py
@@ -1,11 +1,26 @@
-# Avoid importing concrete engines at module import time to prevent circular imports
+from typing import TYPE_CHECKING
+
 from .rollout_engine import ModelOutput, RolloutEngine
+from .types import TinkerTokenInput, TinkerTokenOutput, TokenInput, Tokenizer, TokenOutput, VerlTokenInput, VerlTokenOutput
+
+if TYPE_CHECKING:
+    from .tinker_engine import TinkerEngine
+    from .verl_engine import VerlEngine
 
 __all__ = [
     "ModelOutput",
     "RolloutEngine",
     "OpenAIEngine",
+    "TinkerEngine",
     "VerlEngine",
+    # Token types
+    "TokenInput",
+    "TokenOutput",
+    "TinkerTokenInput",
+    "TinkerTokenOutput",
+    "VerlTokenInput",
+    "VerlTokenOutput",
+    "Tokenizer",
 ]
 
 
@@ -14,6 +29,10 @@ def __getattr__(name):
         from .openai_engine import OpenAIEngine as _OpenAIEngine
 
         return _OpenAIEngine
+    if name == "TinkerEngine":
+        from .tinker_engine import TinkerEngine as _TinkerEngine
+
+        return _TinkerEngine
     if name == "VerlEngine":
         try:
             from .verl_engine import VerlEngine as _VerlEngine
@@ -21,4 +40,4 @@ def __getattr__(name):
             return _VerlEngine
         except Exception:
             raise AttributeError(name) from None
-    raise AttributeError(name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
