+++ rllm/engine/rollout/verl_engine.py
@@ -1,16 +1,19 @@
 import asyncio
 import uuid
+from typing import cast
 
+from omegaconf import DictConfig
+from typing_extensions import override
 from verl.experimental.agent_loop.agent_loop import AgentLoopManager, AsyncLLMServerManager
-from verl.workers.rollout.replica import TokenOutput
 
 from rllm.engine.rollout.rollout_engine import ModelOutput, RolloutEngine
+from rllm.engine.rollout.types import TokenInput, Tokenizer, TokenOutput, VerlTokenOutput
 from rllm.parser import ChatTemplateParser
 from rllm.workflows import TerminationEvent, TerminationReason
 
 
 class VerlEngine(RolloutEngine):
-    def __init__(self, config, rollout_manager, tokenizer, processor=None, **kwargs):
+    def __init__(self, config: DictConfig, rollout_manager: AgentLoopManager, tokenizer: Tokenizer, processor=None, **kwargs):
         self.config = config
 
         if config.actor_rollout_ref.rollout.name not in ["vllm", "sglang"]:
@@ -43,21 +46,35 @@ class VerlEngine(RolloutEngine):
         print(f"train_sampling_params: {self.train_sampling_params}")
         print(f"val_sampling_params: {self.val_sampling_params}")
 
-        self.validate = False  # flag enabled/disabled by AgentWorkflowEngine.execute_tasks_verl
+    @property
+    def supports_token_in_token_out(self) -> bool:
+        return True
 
-    async def get_model_response(self, messages: list[dict], **kwargs) -> ModelOutput:
+    @override
+    async def get_token_output_from_token_input(self, token_input: TokenInput, **kwargs) -> VerlTokenOutput:
+        token_input = cast(list[int], token_input)
+
+        input_length = len(token_input)
         application_id = kwargs.pop("application_id", str(uuid.uuid4()))
-        validate = self.validate or kwargs.pop("validate", False)
         enforce_max_prompt_length = kwargs.pop("enforce_max_prompt_length", True)
 
+        if enforce_max_prompt_length and input_length > self.max_prompt_length:
+            raise TerminationEvent(TerminationReason.MAX_PROMPT_LENGTH_EXCEEDED)
+
+        sampling_params = self.val_sampling_params.copy() if self.is_validation else self.train_sampling_params.copy()
+        sampling_params.update(kwargs)
+        max_tokens = sampling_params.pop("max_tokens", sampling_params.pop("max_new_tokens", self.max_response_length))
+        # starting from verl 0.7.0, we can pass in per-turn max_tokens into the sampling_params
+        sampling_params["max_tokens"] = max_tokens
+
+        token_output = await self.server_manager.generate(request_id=application_id, prompt_ids=token_input, sampling_params=sampling_params)
+        return token_output
+
+    @override
+    async def get_model_response(self, messages: list[dict], **kwargs) -> ModelOutput:
         # these go to the parser
         tools = kwargs.pop("tools", [])
         accumulate_reasoning = kwargs.pop("accumulate_reasoning", self.accumulate_reasoning)
-
-        sampling_params = self.val_sampling_params.copy() if self.validate or validate else self.train_sampling_params.copy()
-        sampling_params.update(kwargs)
-
-        max_tokens = sampling_params.pop("max_tokens", sampling_params.pop("max_new_tokens", self.max_response_length))
 
         prompt = self.chat_parser.parse(messages, add_generation_prompt=True, is_first_msg=True, tools=tools, accumulate_reasoning=accumulate_reasoning)
         request_prompt_ids = self.tokenizer.encode(prompt, add_special_tokens=False)  # list[int]
@@ -73,19 +90,26 @@ class VerlEngine(RolloutEngine):
             multi_modal_inputs = None
             prompt_ids = request_prompt_ids
 
-        prompt_length = len(prompt_ids)
-        if enforce_max_prompt_length and prompt_length > self.max_prompt_length:
-            raise TerminationEvent(TerminationReason.MAX_PROMPT_LENGTH_EXCEEDED)
+        token_output: TokenOutput = await self.get_token_output_from_token_input(token_input=request_prompt_ids, **kwargs)
+        extra_kwargs = dict(prompt_ids=prompt_ids, multi_modal_inputs=multi_modal_inputs)
+        return self.assemble_model_output(token_input=request_prompt_ids, token_output=token_output, **extra_kwargs)
 
-        token_output: TokenOutput = await self.server_manager.generate(request_id=application_id, prompt_ids=request_prompt_ids, image_data=image_data, sampling_params=sampling_params)  # type: ignore
-        completion_ids: list[int] = token_output.token_ids
-        logprobs: list[float] = token_output.log_probs
+    @override
+    def assemble_model_output(self, token_input: TokenInput, token_output: TokenOutput, **kwargs) -> ModelOutput:
+        prompt_ids = kwargs.pop("prompt_ids", None)
+        multi_modal_inputs = kwargs.pop("multi_modal_inputs", None)
+        prompt_length = len(prompt_ids) if prompt_ids is not None else 0
 
-        finish_reason = "stop"
-        if len(completion_ids) >= max_tokens:
-            finish_reason = "length"
-            completion_ids = completion_ids[:max_tokens]
-            logprobs = logprobs[:max_tokens]
+        token_output = cast(VerlTokenOutput, token_output)
+        completion_ids = token_output.token_ids
+        logprobs = token_output.log_probs
+
+        # convert the stop reason from verl back to the standard finish reason TODO(listar2000): check backward-compatibility
+        reason_mapping = {"aborted": "abort", "completed": "stop"}
+        if token_output.stop_reason is not None:
+            finish_reason = reason_mapping.get(token_output.stop_reason, token_output.stop_reason)
+        else:
+            finish_reason = "stop"
 
         completion_text = self.tokenizer.decode(completion_ids, skip_special_tokens=True)
         # TODO: implement parse_completion for the standard parser
