+++ rllm/experimental/rollout/__init__.py
@@ -1,19 +1,20 @@
-from typing import TYPE_CHECKING
-
-from .rollout_engine import ModelOutput, RolloutEngine
-from .types import TinkerTokenInput, TinkerTokenOutput, TokenInput, Tokenizer, TokenOutput, VerlTokenInput, VerlTokenOutput
-
-if TYPE_CHECKING:
-    from .tinker_engine import TinkerEngine
-    from .verl_engine import VerlEngine
+# Backward compatibility: re-export from canonical location
+from rllm.engine.rollout.rollout_engine import ModelOutput, RolloutEngine  # noqa: F401
+from rllm.engine.rollout.types import (  # noqa: F401
+    TinkerTokenInput,
+    TinkerTokenOutput,
+    TokenInput,
+    Tokenizer,
+    TokenOutput,
+    VerlTokenInput,
+    VerlTokenOutput,
+)
 
 __all__ = [
     "ModelOutput",
-    # Rollout engines
     "RolloutEngine",
     "TinkerEngine",
     "VerlEngine",
-    # Token input/output types
     "TokenInput",
     "TokenOutput",
     "TinkerTokenInput",
@@ -25,12 +26,16 @@ __all__ = [
 
 
 def __getattr__(name):
+    # Lazy imports for engines with heavy dependencies
     if name == "TinkerEngine":
-        from .tinker_engine import TinkerEngine as _TinkerEngine
+        from rllm.engine.rollout.tinker_engine import TinkerEngine as _TinkerEngine
 
         return _TinkerEngine
     if name == "VerlEngine":
-        from .verl_engine import VerlEngine as _VerlEngine
+        try:
+            from rllm.engine.rollout.verl_engine import VerlEngine as _VerlEngine
 
-        return _VerlEngine
+            return _VerlEngine
+        except Exception:
+            raise AttributeError(name) from None
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
